requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile .\requirements.in -o .\requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.4
    # via -r ./requirements.in
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
asyncio==4.0.0
    # via -r ./requirements.in
attrs==26.1.0
    # via aiohttp
colorama==0.4.6
    # via tqdm
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.11
    # via yarl
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
numpy==2.4.4
    # via pandas
pandas==3.0.2
    # via -r ./requirements.in
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pydantic==2.12.5
    # via -r ./requirements.in
pydantic-core==2.41.5
    # via pydantic
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.2
    # via -r ./requirements.in
six==1.17.0
    # via python-dateutil
tqdm==4.67.3
    # via -r ./requirements.in
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via pandas
yarl==1.23.0
    # via aiohttp
